* FIX: Fixed convertions size_t to u32 types

+++ code/engine.vc2008/xrServerEntities/associative_vector_inline.h
@@ -115,10 +115,9 @@ IC	typename _associative_vector::size_type _associative_vector::max_size					() 
 }
 
 TEMPLATE_SPECIALIZATION
-//IC	typename _associative_vector::size_type _associative_vector::size						() const
 IC	u32 _associative_vector::size															() const
 {
-	return				(inherited::size());
+	return				u32(inherited::size());
 }
 
 TEMPLATE_SPECIALIZATION
